+++ src/SmartComponents/SystemDetail/InventoryDetail.js
@@ -7,17 +7,15 @@ import { register } from '../../store';
 import { SystemDetailStore } from '../../store/Reducers/SystemDetailStore';
 import { intl } from '../../Utilities/IntlProvider';
 import messages from '../../Messages';
-import { setPageTitle, useFeatureFlag } from '../../Utilities/Hooks';
+import { setPageTitle } from '../../Utilities/Hooks';
 import { InventoryDetailHead, AppInfo, DetailWrapper } from '@redhat-cloud-services/frontend-components/Inventory';
-import { Alert } from '@patternfly/react-core';
+import { Alert, Grid, GridItem, TextContent, Text } from '@patternfly/react-core';
 import { fetchSystemDetailsAction } from '../../store/Actions/Actions';
 import propTypes from 'prop-types';
-import { clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';
-import { featureFlags } from '../../Utilities/constants';
+import { clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';;
 import PatchSetWizard from '../PatchSetWizard/PatchSetWizard';
 
 const InventoryDetail = ({ match }) => {
-    const isPatchSetEnabled = useFeatureFlag(featureFlags.patch_set);
 
     const [patchSetState, setBaselineState] = React.useState({
         isOpen: false,
@@ -30,8 +28,8 @@ const InventoryDetail = ({ match }) => {
         ({ entityDetails }) => entityDetails && entityDetails.entity
     );
 
-    const hasThirdPartyRepo = useSelector(
-        ({ entityDetails }) => entityDetails && entityDetails.hasThirdPartyRepo
+    const { hasThirdPartyRepo, patchSetName } = useSelector(
+        ({ entityDetails }) => entityDetails ?? {}
     );
     const entityId = match.params?.inventoryId;
     useEffect(() => {
@@ -39,7 +37,7 @@ const InventoryDetail = ({ match }) => {
         return () => {
             dispatch(clearNotifications());
         };
-    }, []);
+    }, [patchSetState.shouldRefresh]);
 
     const pageTitle = entityDetails && `${entityDetails.display_name} - ${intl.formatMessage(messages.titlesSystems)}`;
     setPageTitle(pageTitle);
@@ -56,7 +54,7 @@ const InventoryDetail = ({ match }) => {
                 });
             }}
         >
-            {(patchSetState.isOpen && isPatchSetEnabled) &&
+            {(patchSetState.isOpen) &&
                 <PatchSetWizard systemsIDs={patchSetState.systemsIDs} setBaselineState={setBaselineState} />}
             <Header
                 title=""
@@ -73,18 +71,32 @@ const InventoryDetail = ({ match }) => {
                     }
                 ]}
             >
-                <InventoryDetailHead hideBack actions={isPatchSetEnabled && [
-                    {
-                        title: intl.formatMessage(messages.titlesPatchSetAssignMultipleButton),
-                        key: 'assign-to-patch-set',
-                        onClick: showBaselineModal
-                    }]}
+                <InventoryDetailHead hideBack
+                    showTags
+                    actions={[
+                        {
+                            title: intl.formatMessage(messages.titlesPatchSetAssignMultipleButton),
+                            key: 'assign-to-patch-set',
+                            onClick: showBaselineModal
+                        }]}
                 >
-                    { hasThirdPartyRepo &&
-                        (<Alert className='pf-u-mt-md' isInline variant="info"
-                            title={intl.formatMessage(messages.textThirdPartyInfo)}>
-                        </Alert>)
-                    }
+                    <Grid>
+                        <GridItem>
+                            {patchSetName && <TextContent>
+                                <Text>
+                                    {`${intl.formatMessage(messages.labelsColumnsPatchSet)}: ${patchSetName}`}
+                                </Text>
+                            </TextContent>
+                            }
+                        </GridItem>
+                        <GridItem>
+                            {hasThirdPartyRepo &&
+                                (<Alert className='pf-u-mt-md' isInline variant="info"
+                                    title={intl.formatMessage(messages.textThirdPartyInfo)}>
+                                </Alert>)
+                            }
+                        </GridItem>
+                    </Grid>
                 </InventoryDetailHead>
             </Header>
             <Main>