+++ src/Utilities/DataMappers.js
@@ -194,7 +194,7 @@ export const createPackageSystemsRows = (rows, selectedRows = {}) => {
 };
 
 export const createSystemPackagesRows = (rows, selectedRows = {}) => {
-    if (rows.length !== 0) {
+    if (rows && rows.length !== 0) {
         return rows.map(pkg => {
             const pkgNEVRA = `${pkg.name}-${pkg.evra}`;
             const pkgUpdates = pkg.updates || [];
@@ -316,6 +316,9 @@ export const createSystemsRowsReview = (rows, selectedRows) => {
                     },
                     {
                         title: attributes.os
+                    },
+                    {
+                        title: attributes.baseline_name
                     }
                 ]
             };